requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.3.3
automium==0.2.6
automium-web==0.1.1
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
crypto==1.4.1
cvxpy==1.0.28
decorator==4.4.2
dill==0.3.1.1
distlib==0.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.8
django-extensions==3.1.0
django-nine==0.2.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
ecos==2.0.7.post1
elasticsearch==7.7.0
elasticsearch-dsl==7.2.0
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
httpie==2.0.0
idna==2.9
importlib-metadata==2.0.0
ipython==7.19.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
legacy==0.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
mod-wsgi==4.7.1
multiprocess==0.70.9
mysql-connector==2.2.9
Naked==0.1.31
numpy==1.19.1
osqp==0.6.1
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.11
pycparser==2.20
pycryptodomex==3.9.8
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.23.0
scipy==1.4.1
scs==2.1.1.post2
sgmllib3k==1.0.0
shellescape==3.8.1
six==1.14.0
sqlparse==0.4.1
traitlets==5.0.5
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.0.32
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.11.2
xlrd==1.2.0
zipp==3.2.0
